extract account slice
