+++ app/storage/index.ts
@@ -1,54 +1,29 @@
-import { getAddress } from "ethers"
 import { applyPatches, type Patch } from "immer"
-import { v4 as uuidV4 } from "uuid"
 import browser from "webextension-polyfill"
 import { create } from "zustand"
 import { useShallow } from "zustand/react/shallow"
 
 import { StorageAction } from "~background/messages/storage"
-import { PasskeyAccount } from "~packages/account/PasskeyAccount"
 import { type UserOperation } from "~packages/bundler/userOperation"
 import address from "~packages/util/address"
-import number from "~packages/util/number"
 import {
   TransactionStatus,
   TransactionType,
   type ERC4337TransactionRejected,
-  type ERC4337TransactionSent,
-  type Token
+  type ERC4337TransactionSent
 } from "~storage/local/state"
-import type { BigNumberish, HexString } from "~typing"
+import type { HexString } from "~typing"
 
 import { StorageMessenger } from "./messenger"
 import { background } from "./middleware/background"
+import { createAccountSlice, type AccountSlice } from "./slice/account"
 import { createProfileSlice, type ProfileSlice } from "./slice/profile"
 import { createStateSlice, type StateSlice } from "./slice/state"
 
 const storageMessenger = new StorageMessenger()
 
 // TODO: Split as slices
-interface Storage extends StateSlice, ProfileSlice {
-  /* Account */
-
-  createAccount: (account: PasskeyAccount, networkId: string) => Promise<void>
-
-  switchAccount: (accountId: string) => Promise<void>
-
-  updateBalance: (accountId: string, balance: BigNumberish) => Promise<void>
-
-  importToken: (accountId: string, token: Token) => Promise<void>
-
-  updateToken: (
-    accountId: string,
-    tokenAddress: HexString,
-    update: {
-      balance?: BigNumberish
-      symbol?: string
-    }
-  ) => Promise<void>
-
-  removeToken: (accountId: string, tokenAddress: HexString) => Promise<void>
-
+interface Storage extends StateSlice, AccountSlice, ProfileSlice {
   /* Network */
 
   switchNetwork: (networkId: string) => Promise<void>
@@ -84,96 +59,8 @@ export const useStorage = create<Storage>()(
   background(
     (set, get, ...others) => ({
       ...createStateSlice(set, get, ...others),
+      ...createAccountSlice(set, get, ...others),
       ...createProfileSlice(set, get, ...others),
-
-      /* Account */
-
-      createAccount: async (account: PasskeyAccount, networkId: string) => {
-        const id = uuidV4()
-        const data = account.dump()
-        await set(({ state }) => {
-          const network = state.network[networkId]
-          state.account[id] = {
-            ...data,
-            chainId: network.chainId,
-            // TODO: Design a value object
-            publicKey: {
-              x: number.toHex(data.publicKey.x),
-              y: number.toHex(data.publicKey.y)
-            },
-            salt: number.toHex(data.salt),
-            // TODO: Design an account periphery prototype
-            transactionLog: {},
-            balance: "0x00",
-            tokens: []
-          }
-          // Set the new account as active
-          network.accountActive = id
-        })
-      },
-
-      switchAccount: async (accountId: string) => {
-        await set(({ state }) => {
-          const { account, network, networkActive } = state
-          if (account[accountId].chainId !== network[networkActive].chainId) {
-            throw new Error("Cannot switch to account in other network")
-          }
-          state.network[state.networkActive].accountActive = accountId
-        })
-      },
-
-      updateBalance: async (accountId: string, balance: BigNumberish) => {
-        await set(({ state }) => {
-          state.account[accountId].balance = number.toHex(balance)
-        })
-      },
-
-      importToken: async (accountId: string, token: Token) => {
-        await set(({ state }) => {
-          state.account[accountId].tokens.push({
-            address: getAddress(token.address),
-            symbol: token.symbol,
-            decimals: token.decimals,
-            balance: token.balance
-          })
-        })
-      },
-
-      removeToken: async (accountId: string, tokenAddress: HexString) => {
-        await set(({ state }) => {
-          const tokenIndex = state.account[accountId].tokens.findIndex(
-            (token) => address.isEqual(token.address, tokenAddress)
-          )
-          if (tokenIndex < 0) {
-            throw new Error(`Unknown token: ${tokenAddress}`)
-          }
-          state.account[accountId].tokens.splice(tokenIndex, 1)
-        })
-      },
-
-      updateToken: async (
-        accountId: string,
-        tokenAddress: HexString,
-        update: {
-          balance?: BigNumberish
-          symbol?: string
-        }
-      ) => {
-        await set(({ state }) => {
-          const token = state.account[accountId].tokens.find((token) =>
-            address.isEqual(token.address, tokenAddress)
-          )
-          if (!token) {
-            throw new Error(`Unknown token: ${tokenAddress}`)
-          }
-          if (update.balance) {
-            token.balance = number.toHex(update.balance)
-          }
-          if (update.symbol) {
-            token.symbol = update.symbol
-          }
-        })
-      },
 
       /* Network */
